# api/api-contract.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Northstar Dashboard API",
    "version": "1.0.0",
    "description": "Backend contract for the Northstar dashboard, project roadmaps, versions, GitHub issue-aligned todos, and WIP assignment workflows."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development backend"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/projects": {
      "get": {
        "tags": ["Projects"],
        "summary": "List projects for the sidebar",
        "operationId": "listProjects",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A page of projects.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectList"
                },
                "example": {
                  "items": [
                    {
                      "id": "northstar-web",
                      "companyId": "northstar",
                      "name": "Northstar Web",
                      "path": "/northstar-web",
                      "versionCount": 2,
                      "wipTodoCount": 2,
                      "createdAt": "2026-01-15T09:00:00.000Z",
                      "updatedAt": "2026-07-27T11:48:00.000Z"
                    }
                  ],
                  "nextCursor": null
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Projects"],
        "summary": "Create a project",
        "operationId": "createProject",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectRequest"
              },
              "example": {
                "companyId": "northstar",
                "name": "Mobile Client",
                "path": "/mobile-client"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The created project.",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string"
                },
                "example": "/api/v1/projects/mobile-client"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      }
    },
    "/api/v1/projects/{projectId}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/ProjectId"
        }
      ],
      "get": {
        "tags": ["Projects"],
        "summary": "Get a project roadmap",
        "operationId": "getProject",
        "responses": {
          "200": {
            "description": "The project with versions and WIP todos.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                },
                "example": {
                  "id": "northstar-web",
                  "companyId": "northstar",
                  "name": "Northstar Web",
                  "path": "/northstar-web",
                  "versions": [
                    {
                      "id": "northstar-v2-2",
                      "projectId": "northstar-web",
                      "name": "v2.2",
                      "summary": "Dashboard workflows and reporting improvements.",
                      "status": "in-progress",
                      "todos": [
                        {
                          "id": "todo-dashboard-shell",
                          "projectId": "northstar-web",
                          "versionId": "northstar-v2-2",
                          "issueNumber": 142,
                          "issueUrl": "https://github.com/northstar/web/issues/142",
                          "pullRequestNumber": 316,
                          "pullRequestUrl": "https://github.com/northstar/web/pull/316",
                          "title": "Ship the responsive dashboard shell",
                          "description": "Finish the collapsible application sidebar, top bar, and full-width routed content layout.",
                          "status": "done",
                          "isMerged": false,
                          "mergedAt": null,
                          "createdAt": "2026-07-20T08:30:00.000Z",
                          "updatedAt": "2026-07-27T10:15:00.000Z"
                        }
                      ],
                      "createdAt": "2026-07-01T09:00:00.000Z",
                      "updatedAt": "2026-07-27T10:15:00.000Z",
                      "releasedAt": null
                    }
                  ],
                  "wipTodos": [
                    {
                      "id": "todo-github-sync-boundaries",
                      "projectId": "northstar-web",
                      "versionId": null,
                      "issueNumber": 145,
                      "issueUrl": "https://github.com/northstar/web/issues/145",
                      "pullRequestNumber": null,
                      "pullRequestUrl": null,
                      "title": "Define GitHub synchronization boundaries",
                      "description": "Document which issue fields are managed by the roadmap and how conflicts with GitHub updates are resolved.",
                      "status": "draft",
                      "isMerged": false,
                      "mergedAt": null,
                      "createdAt": "2026-07-26T12:00:00.000Z",
                      "updatedAt": "2026-07-26T12:00:00.000Z"
                    }
                  ],
                  "createdAt": "2026-01-15T09:00:00.000Z",
                  "updatedAt": "2026-07-27T11:48:00.000Z"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "patch": {
        "tags": ["Projects"],
        "summary": "Update project metadata",
        "operationId": "updateProject",
        "x-ui-usage": "The project settings page uses this endpoint to update the project name and folder path.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated project.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "tags": ["Projects"],
        "summary": "Delete a project and its roadmap",
        "operationId": "deleteProject",
        "x-ui-usage": "Supporting CRUD operation; no delete control is currently exposed.",
        "responses": {
          "204": {
            "description": "The project was deleted."
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/v1/projects/{projectId}/versions": {
      "parameters": [
        {
          "$ref": "#/components/parameters/ProjectId"
        }
      ],
      "post": {
        "tags": ["Versions"],
        "summary": "Add a version to a project",
        "operationId": "createVersion",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVersionRequest"
              },
              "example": {
                "name": "v2.3",
                "summary": "GitHub synchronization and roadmap permissions."
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The created pending version.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectVersion"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      }
    },
    "/api/v1/projects/{projectId}/versions/{versionId}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/ProjectId"
        },
        {
          "$ref": "#/components/parameters/VersionId"
        }
      ],
      "get": {
        "tags": ["Versions"],
        "summary": "Get one version with its todos",
        "operationId": "getVersion",
        "responses": {
          "200": {
            "description": "The requested version.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectVersion"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "patch": {
        "tags": ["Versions"],
        "summary": "Update version metadata or workflow status",
        "operationId": "updateVersion",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVersionRequest"
              },
              "example": {
                "status": "complete"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated version.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectVersion"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      },
      "delete": {
        "tags": ["Versions"],
        "summary": "Delete a version",
        "operationId": "deleteVersion",
        "x-ui-usage": "The version card exposes this action only for pending versions and requires confirmation.",
        "responses": {
          "204": {
            "description": "The version was deleted."
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      }
    },
    "/api/v1/projects/{projectId}/versions/{versionId}/release": {
      "parameters": [
        {
          "$ref": "#/components/parameters/ProjectId"
        },
        {
          "$ref": "#/components/parameters/VersionId"
        }
      ],
      "post": {
        "tags": ["Versions"],
        "summary": "Release a completed version",
        "operationId": "releaseVersion",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReleaseVersionRequest"
              },
              "example": {
                "releaseNotes": "Accessibility and performance hardening."
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The released version.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectVersion"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      }
    },
    "/api/v1/projects/{projectId}/versions/{versionId}/todos": {
      "parameters": [
        {
          "$ref": "#/components/parameters/ProjectId"
        },
        {
          "$ref": "#/components/parameters/VersionId"
        }
      ],
      "post": {
        "tags": ["Todos"],
        "summary": "Create a todo in a version",
        "operationId": "createVersionTodo",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVersionTodoRequest"
              },
              "example": {
                "title": "Add synchronization status indicator",
                "description": "Show whether roadmap changes are saved locally, syncing, or fully synchronized."
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The created draft todo and associated issue metadata.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectTodo"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      }
    },
    "/api/v1/projects/{projectId}/wip/todos": {
      "parameters": [
        {
          "$ref": "#/components/parameters/ProjectId"
        }
      ],
      "post": {
        "tags": ["Todos"],
        "summary": "Create an unassigned draft todo in WIP",
        "operationId": "createWipTodo",
        "x-ui-usage": "The WIP column uses this endpoint to create unassigned draft todos.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWipTodoRequest"
              },
              "example": {
                "title": "Define GitHub synchronization boundaries",
                "description": "Document which issue fields are managed by the roadmap and how conflicts are resolved."
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The created draft todo.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectTodo"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/v1/projects/{projectId}/todos/{todoId}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/ProjectId"
        },
        {
          "$ref": "#/components/parameters/TodoId"
        }
      ],
      "get": {
        "tags": ["Todos"],
        "summary": "Get todo details",
        "operationId": "getTodo",
        "responses": {
          "200": {
            "description": "The requested todo.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectTodo"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "patch": {
        "tags": ["Todos"],
        "summary": "Update todo details or status",
        "operationId": "updateTodo",
        "x-ui-usage": "The todo item uses this endpoint to mark an assigned draft todo as planned. That transition creates its branch.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTodoRequest"
              },
              "example": {
                "status": "done"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated todo.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectTodo"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      },
      "delete": {
        "tags": ["Todos"],
        "summary": "Delete a todo",
        "operationId": "deleteTodo",
        "x-ui-usage": "Todo items expose this action only for draft todos and require confirmation.",
        "responses": {
          "204": {
            "description": "The todo was deleted."
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      }
    },
    "/api/v1/projects/{projectId}/todos/{todoId}/assign": {
      "parameters": [
        {
          "$ref": "#/components/parameters/ProjectId"
        },
        {
          "$ref": "#/components/parameters/TodoId"
        }
      ],
      "post": {
        "tags": ["Todos"],
        "summary": "Assign a WIP todo to a version",
        "operationId": "assignTodoToVersion",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignTodoRequest"
              },
              "example": {
                "versionId": "northstar-v2-2"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The assigned todo. Its status is changed from draft to planned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectTodo"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      }
    },
    "/api/v1/projects/{projectId}/todos/{todoId}/merge": {
      "parameters": [
        {
          "$ref": "#/components/parameters/ProjectId"
        },
        {
          "$ref": "#/components/parameters/TodoId"
        }
      ],
      "post": {
        "tags": ["Todos"],
        "summary": "Mark a completed todo as merged",
        "operationId": "mergeTodo",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MergeTodoRequest"
              },
              "example": {
                "mergeCommitSha": "f2d85c18d3a984706bdc7a875c2386dd1428cb24"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The merged todo.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectTodo"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
...
      "ProjectId": {
        "name": "projectId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "minLength": 1
        },
        "example": "northstar-web"
      },
      "VersionId": {
        "name": "versionId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "minLength": 1
        },
        "example": "northstar-v2-2"
      },
...
    },
    "schemas": {
...
      "ProjectList": {
        "type": "object",
        "additionalProperties": false,
        "required": ["items", "nextCursor"],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProjectSummary"
            }
          },
          "nextCursor": {
            "type": ["string", "null"]
          }
        }
      },
      "Project": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "companyId",
          "name",
          "path",
          "versions",
          "wipTodos",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "companyId": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "path": {
            "type": "string",
            "minLength": 1,
            "description": "An opaque display path selected by the user. It does not grant server filesystem access."
          },
          "versions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProjectVersion"
            }
          },
          "wipTodos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProjectTodo"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ProjectVersion": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "projectId",
          "name",
          "summary",
          "status",
          "todos",
          "createdAt",
          "updatedAt",
          "releasedAt"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "projectId": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 60
          },
          "summary": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1000
          },
          "status": {
            "$ref": "#/components/schemas/VersionStatus"
          },
          "todos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProjectTodo"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "releasedAt": {
            "type": ["string", "null"],
            "format": "date-time"
          }
        }
      },
...
      "CreateProjectRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["companyId", "name", "path"],
        "properties": {
          "companyId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "path": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2048
          }
        }
      },
      "UpdateProjectRequest": {
        "type": "object",
        "additionalProperties": false,
        "minProperties": 1,
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "path": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2048
          }
        }
      },
      "CreateVersionRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["name", "summary"],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 60
          },
          "summary": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1000
          }
        }
      },
      "UpdateVersionRequest": {
        "type": "object",
        "additionalProperties": false,
        "minProperties": 1,
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 60
          },
          "summary": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1000
          },
          "status": {
            "type": "string",
            "enum": ["pending", "ready", "in-progress", "complete"],
            "description": "Use the release endpoint for the released state."
          }
        }
      },
...
    },
    "responses": {
...
      "ValidationError": {
        "description": "The request failed validation.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": {
                "code": "VALIDATION_ERROR",
                "message": "The request contains invalid fields.",
                "details": [
                  {
                    "field": "name",
                    "message": "Name is required."
                  }
                ],
                "requestId": "req_01J3X8Y7P5"
              }
            }
          }
        }
      },
      "NotFound": {
        "description": "The requested resource was not found.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": {
                "code": "NOT_FOUND",
                "message": "The requested resource was not found.",
                "requestId": "req_01J3X8Y7P6"
              }
            }
          }
        }
      },
      "Conflict": {
        "description": "The operation conflicts with the resource's current state.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": {
                "code": "INVALID_STATE_TRANSITION",
                "message": "The resource cannot perform this operation in its current state.",
                "requestId": "req_01J3X8Y7P7"
              }
            }
          }
        }
      },
...
    }
  },
...
}